Split the `pyi` stubs up and add more typing information
