+++ pyvips/base.pyi
@@ -1,0 +1,16 @@
+from collections.abc import Callable
+
+def leak_set(leak: bool) -> None: ...
+def shutdown() -> None: ...
+def version(flag: int) -> int: ...
+def get_suffixes() -> list[str]: ...
+def at_least_libvips(x: int, y: int) -> bool: ...
+def type_find(basename: str, nickname: str) -> int: ...
+def type_name(gtype: int) -> str: ...
+def nickname_find(gtype: int) -> str: ...
+def type_from_name(name: str) -> int: ...
+def type_map(gtype: int, fn: Callable[..., int]) -> int: ...
+def values_for_enum(gtype: int) -> list[str]: ...
+def values_for_flag(gtype: int) -> list[str]: ...
+def enum_dict(gtype: int) -> dict[str, int]: ...
+def flags_dict(gtype: int) -> dict[str, int]: ...
