+++ pyvips/enums.pyi
