+++ pyvips/vobject.pyi
@@ -1,0 +1,22 @@
+import logging
+from typing import Any, ClassVar
+
+from pyvips.gobject import GObject
+
+logger: logging.Logger
+
+class VipsObject(GObject):
+    vobject: Any
+    gobject: Any
+    _pspec_cache: ClassVar[dict[Any, dict[str, Any]]]
+
+    def __init__(self, pointer: Any) -> None: ...
+    @staticmethod
+    def print_all() -> None: ...
+    def _get_pspec(self, name: str) -> Any | None: ...
+    def get_typeof(self, name: str) -> int: ...
+    def get_blurb(self, name: str) -> str: ...
+    def get(self, name: str) -> Any: ...
+    def set(self, name: str, value: Any) -> None: ...
+    def set_string(self, string_options: str) -> bool: ...
+    def get_description(self) -> str: ...
