+++ pyvips/vregion.pyi
@@ -1,0 +1,12 @@
+from typing import Any
+
+from pyvips.vimage import Image
+from pyvips.vobject import VipsObject
+
+class Region(VipsObject):
+    def __init__(self, pointer: Any) -> None: ...
+    @staticmethod
+    def new(image: Image) -> Region: ...
+    def width(self) -> int: ...
+    def height(self) -> int: ...
+    def fetch(self, x: int, y: int, w: int, h: int) -> memoryview: ...
