+++ pyvips/vtargetcustom.pyi
@@ -1,0 +1,16 @@
+import logging
+from typing import Callable
+
+from pyvips.vtarget import Target
+
+logger: logging.Logger
+
+_BufferLike = bytes | bytearray | memoryview
+
+class TargetCustom(Target):
+    def __init__(self) -> None: ...
+    def on_write(self, handler: Callable[[_BufferLike], int]) -> None: ...
+    def on_read(self, handler: Callable[[int], _BufferLike | None]) -> None: ...
+    def on_seek(self, handler: Callable[[int, int], int]) -> None: ...
+    def on_end(self, handler: Callable[[], int]) -> None: ...
+    def on_finish(self, handler: Callable[[], int]) -> None: ...
